平复了心情~ -调整了代码架构 -加入了setting功能

+++ lib/ESP32_menu_v1/1_driver/keyboard/keyboard_struct.h
@@ -29,7 +29,7 @@ typedef struct menu_keyboard_t{
 
     uint8_t key_enum;
     uint8_t key_value;
-    uint8_t press_time;
+    uint16_t press_time;
 
 }menu_keyboard;
